pyproject.toml
[tool.poetry]
name = "polus-pipelines"
version = "0.1.0"
description = "Build and execute pipelines of polus plugins on Compute."
authors = ["Antoine Gerardin <antoine.gerardin@nih.gv>"]
readme = "README.md"
packages = [{include = "polus", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
polus-plugins = {git = "https://github.com/PolusAI/polus-plugins"}
polus-render = "^0.0.4.0.1.5"
workflow-inference-compiler = {git = "https://github.com/PolusAI/workflow-inference-compiler.git"}
load-dotenv = "^0.1.0"
typer = "^0.9.0"
pyyaml = "^6.0.1"
solara = "^1.25.1"
cwl-utils="0.31"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
nbconvert = "^7.11.0"
pytest = "^7.4.4"
bump2version = "^1.0.1"
pre-commit = "^3.3.3"
black = "^23.3.0"
ruff = "^0.0.274"
mypy = "^1.4.0"
pytest-xdist = "^3.3.1"
pytest-sugar = "^0.9.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
